Split up file structure
