+++ backend/love_api/api/models/todo_calendar_model.py
@@ -3,7 +3,8 @@ from django.db import models
 from django.utils.timezone import now
 
 class TodoCalendar(models.Model):
-    user = models.ForeignKey(User, related_name='calendar', on_delete=models.CASCADE)
+    user = models.ForeignKey(User, related_name='calendar', 
+    	on_delete=models.CASCADE)
     message = models.TextField(max_length=800, blank=False, null=False)
     status = models.CharField(max_length=10, default='full')
     date = models.DateTimeField(blank=False, null=False)
@@ -12,4 +13,5 @@ class TodoCalendar(models.Model):
     	return now() > self.event_date
 
     def __str__(self):
-        return "id: {}, username: {}".format(self.user.id, self.user.username)
+        return "id: {}, username: {}".format(
+        	self.user.id, self.user.username)
