+++ backend/love_api/api/serializers/important_number_serializer.py
@@ -1,0 +1,17 @@
+from django.contrib.auth.models import User, Group
+from rest_framework import serializers
+from .user_serializer import *
+from api.models import *
+
+class ImportantNumberSerializer(serializers.ModelSerializer):
+    """
+    Serializer to map the Model instance into JSON format.
+    """
+    user = serializers.ReadOnlyField(source='user.id')
+
+    class Meta:
+        """
+        Meta class to map serializer's fields with the model fields.
+        """
+        model = ImportantNumber
+        exclude = ()
