+++ backend/love_api/api/serializers/todo_calendar_serializer.py
@@ -1,0 +1,17 @@
+from django.contrib.auth.models import User, Group
+from rest_framework import serializers
+from .user_serializer import *
+from api.models import *
+
+class TodoCalendarSerializer(serializers.ModelSerializer):
+    """
+    Serializer to map the Model instance into JSON format.
+    """
+    user = serializers.ReadOnlyField(source='user.id')
+
+    class Meta:
+        """
+        Meta class to map serializer's fields with the model fields.
+        """
+        model = TodoCalendar
+        exclude = ()
