+++ backend/love_api/api/views/__init__.py
@@ -1,1 +1,10 @@
-from .views import *
+from .important_number_viewset import *
+from .message_viewset import *
+from .notification_viewset import *
+from .photo_viewset import *
+from .todo_calendar_viewset import *
+from .user_viewset import *
+from .group_viewset import *
+from .relationship_viewset import *
+from .profile_viewset import *
+from .limited_viewset import *
